+++ src/components/Login.tsx
@@ -1,0 +1,16 @@
+import { useAuth0 } from "@auth0/auth0-react";
+import { Button, Text } from "@chakra-ui/react";
+
+const Login = () => {
+  const { user, isAuthenticated, loginWithRedirect, logout } = useAuth0();
+
+  return (
+    <>
+      <Text>お疲れ様です！</Text>
+      <Text>ログインしてから打刻してください _(._.)_</Text>
+      <Button onClick={loginWithRedirect}>Log in</Button>
+    </>
+  );
+};
+
+export { Login };
